+++ auraed/src/runtime/cells/cell.rs
@@ -28,19 +28,62 @@
  *                                                                            *
 \* -------------------------------------------------------------------------- */
 
-use crate::runtime::cells::executable::ExecutableError;
-use crate::runtime::cells::{
-    validation::ValidatedCell, CellName, Executable, ExecutableName,
+use super::{
+    validation::ValidatedCell, CellName, Executable, ExecutableError,
+    ExecutableName,
 };
 use cgroups_rs::cgroup_builder::CgroupBuilder;
 use cgroups_rs::{hierarchies, Cgroup, Hierarchy};
-use log::{error, info};
+use log::error;
+use log::info;
 use std::collections::HashMap;
 use std::process::{Command, ExitStatus};
 use thiserror::Error;
 use tonic::Status;
 
-type Result<T> = std::result::Result<T, CellError>;
+pub(crate) type Result<T> = std::result::Result<T, CellError>;
+
+#[derive(Error, Debug)]
+pub(crate) enum CellError {
+    #[error("cell '{cell_name}' already exists'")]
+    CellExists { cell_name: CellName },
+    #[error("cell '{cell_name}' not found'")]
+    CellNotFound { cell_name: CellName },
+    #[error("cell '{cell_name}' could not be freed: {source}")]
+    FailedToFreeCell { cell_name: CellName, source: cgroups_rs::error::Error },
+    #[error(
+        "cell '{cell_name}' already has an executable '{executable_name}'"
+    )]
+    ExecutableExists { cell_name: CellName, executable_name: ExecutableName },
+    #[error("cell '{cell_name} could not find executable '{executable_name}'")]
+    ExecutableNotFound { cell_name: CellName, executable_name: ExecutableName },
+    #[error("cell '{cell_name}': {source}")]
+    ExecutableError { cell_name: CellName, source: ExecutableError },
+    #[error(
+        "cell '{cell_name}' failed to add executable (executable:?): {source}"
+    )]
+    FailedToAddExecutable {
+        cell_name: CellName,
+        executable: Executable,
+        source: cgroups_rs::error::Error,
+    },
+}
+
+impl From<CellError> for Status {
+    fn from(err: CellError) -> Self {
+        let msg = err.to_string();
+        error!("{msg}");
+        match err {
+            CellError::CellExists { .. }
+            | CellError::ExecutableExists { .. } => Status::already_exists(msg),
+            CellError::CellNotFound { .. }
+            | CellError::ExecutableNotFound { .. } => Status::not_found(msg),
+            CellError::ExecutableError { .. }
+            | CellError::FailedToFreeCell { .. }
+            | CellError::FailedToAddExecutable { .. } => Status::internal(msg),
+        }
+    }
+}
 
 #[derive(Debug)]
 pub(crate) struct Cell {
@@ -50,7 +93,6 @@ pub(crate) struct Cell {
 }
 
 impl Cell {
-    // TODO: This fn signature ties cells module to runtime module (refactor to better solution)
     // Here is where we define the "default" cgroup parameters for Aurae cells
     pub fn allocate(cell_spec: ValidatedCell) -> Self {
         let ValidatedCell { name, cpu_cpus, cpu_shares, cpu_mems, cpu_quota } =
@@ -168,44 +210,4 @@ fn hierarchy() -> Box<dyn Hierarchy> {
     // hierarchies::auto() // Uncomment to auto detect Cgroup hierarchy
     // hierarchies::V2
     Box::new(hierarchies::V2::new())
-}
-
-#[derive(Error, Debug)]
-pub(crate) enum CellError {
-    #[error("cell '{cell_name}' already exists'")]
-    CellExists { cell_name: CellName },
-    #[error("cell '{cell_name}' not found'")]
-    CellNotFound { cell_name: CellName },
-    #[error("cell '{cell_name}' could not be freed: {source}")]
-    FailedToFreeCell { cell_name: CellName, source: cgroups_rs::error::Error },
-    #[error(
-        "cell '{cell_name}' already has an executable '{executable_name}'"
-    )]
-    ExecutableExists { cell_name: CellName, executable_name: ExecutableName },
-    #[error("cell '{cell_name} could not find executable '{executable_name}'")]
-    ExecutableNotFound { cell_name: CellName, executable_name: ExecutableName },
-    #[error("cell '{cell_name}': {source}")]
-    ExecutableError { cell_name: CellName, source: ExecutableError },
-    #[error("cell '{cell_name}' failed to add executable (executable:?): {source}")]
-    FailedToAddExecutable {
-        cell_name: CellName,
-        executable: Executable,
-        source: cgroups_rs::error::Error,
-    },
-}
-
-impl From<CellError> for Status {
-    fn from(err: CellError) -> Self {
-        let msg = err.to_string();
-        error!("{msg}");
-        match err {
-            CellError::CellExists { .. }
-            | CellError::ExecutableExists { .. } => Status::already_exists(msg),
-            CellError::CellNotFound { .. }
-            | CellError::ExecutableNotFound { .. } => Status::not_found(msg),
-            CellError::ExecutableError { .. }
-            | CellError::FailedToFreeCell { .. }
-            | CellError::FailedToAddExecutable { .. } => Status::internal(msg),
-        }
-    }
 }
