+++ auraed/src/runtime/cells/cells_table.rs
@@ -28,15 +28,13 @@
  *                                                                            *
 \* -------------------------------------------------------------------------- */
 
-use crate::runtime::cells::cell::CellError;
-use crate::runtime::cells::CellServiceError;
-use crate::runtime::cells::{Cell, CellName, Result};
+use super::{Cell, CellError, CellName, CellServiceError, Result};
 use std::{
     collections::HashMap,
     sync::{Arc, Mutex},
 };
 
-/// CgroupTable is the in-memory store for the list of cgroups created with Aurae.
+/// CellsTable is the in-memory store for the list of cells created with Aurae.
 #[derive(Debug, Default, Clone)]
 pub(crate) struct CellsTable {
     // TODO (future-highway): would a RWLock be more performant?
@@ -50,9 +48,7 @@ pub(crate) struct CellsTable {
 // - Get Cgroup and pids from executable_name
 
 impl CellsTable {
-    /// Add the [cgroup] to the cache with key [cell_name].
-    /// Note that this does not take ownership of the cgroup and instead clones it.
-    /// The clone can be retrieved once it's removed from the cache.
+    /// Add the [cell] to the cache with key [cell_name].
     /// Returns an error if a duplicate [cell_name] already exists in the cache.
     pub(crate) fn insert(&self, cell_name: CellName, cell: Cell) -> Result<()> {
         let mut cache = self
@@ -80,18 +76,6 @@ impl CellsTable {
 
         Ok(cache.contains_key(cell_name))
     }
-
-    // /// Return a clone of the cgroup keyed by [cell_name] from the cache or None if it is not found.
-    // /// Does not relinquish ownership.
-    // /// Returns an error if we fail to lock the cache.
-    // pub(crate) fn get(&self, cell_name: &CellName) -> Result<Option<Cgroup>> {
-    //     let cache = self
-    //         .cache
-    //         .lock()
-    //         .map_err(|e| anyhow!("failed to lock cgroup_table: {e:?}"))?;
-    //     let cgroup = cache.get(cell_name).cloned();
-    //     Ok(cgroup)
-    // }
 
     pub(crate) fn get_then<F, R>(&self, cell_name: &CellName, f: F) -> Result<R>
     where
