+++ auraed/src/runtime/cells/mod.rs
@@ -1,16 +1,15 @@
 use cell::{Cell, CellError};
 use cell_name::CellName;
 pub(crate) use cell_service::CellService;
+use cell_service::{CellServiceError, Result};
 use cells_table::CellsTable;
-use error::{CellServiceError, Result};
-use executable::Executable;
+use executable::{Executable, ExecutableError};
 use executable_name::ExecutableName;
 
 mod cell;
 mod cell_name;
 mod cell_service;
 mod cells_table;
-mod error;
 mod executable;
 mod executable_name;
 mod validation;
